rename `types.ts` to `tokenTypes.ts`
